Move identity to core util
