--- core/identity.js
@@ -1,6 +1,0 @@
-'use strict'
-
-module.exports =
-function identity(result){
-  return result
-}
