+++ core/transformer.js
@@ -4,8 +4,8 @@ var /* global Symbol */
     symbolExists = typeof Symbol !== 'undefined',
     symbol = symbolExists ? Symbol('transformer') : '@@transformer',
     completing = require('./completing'),
-    identity = require('./identity'),
     util = require('./util'),
+    identity = util.identity,
     isArray = util.isArray,
     isFunction = util.isFunction,
     isString = util.isString,
