+++ test/util.js
@@ -34,7 +34,7 @@ test('isReduced', function(t){
 
 test('is', function(t){
   t.ok(tp.isFunction(test))
-  t.ok(tp.isFunction(tr.identity))
+  t.ok(tp.isFunction(tp.identity))
   t.ok(tp.isString(''))
   t.ok(tp.isRegExp(/./))
   t.ok(tp.isUndefined())
